backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
deepface
numpy
opencv-python-headless
